FIX: код-ревью — XSS, утечки данных, dead code, слабые тесты
Security:
- XSS: добавить escapeHtml() в dashboard JS (users, events, chats)
- ApiExceptionHandler: не раскрывать внутренние ошибки клиентам
- application.properties: убрать дефолтные пароли admin/user
- BotUserAccessPolicy: Long.longValue() вместо == для RBAC
- .gitignore: добавить *.db, *.pem, *.key, *.session

Bugs:
- get_session.py: добавить f-prefix в print
- queue_consumer.py: datetime.utcnow() → datetime.now(timezone.utc)
- CountingOutputStream: добавить close() delegation
- build.yml: порт health check 8081 → 8080
- java_client.py: логировать ошибки в _transform_entities и update_queue_position

Dead code:
- Удалить FileConversionService (не используется)
- Удалить TelegramExporterInterface (не имплементирован)

Tests:
- Усилить assertions: isNotNull → contains(expected_value)

Docs:
- SETUP.md: исправить описание порта prod

+++ src/main/java/com/tcleaner/dashboard/security/BotUserAccessPolicy.java
@@ -44,6 +44,6 @@ public class BotUserAccessPolicy {
         if (role == DashboardRole.ADMIN) {
             return true;
         }
-        return ownBotUserId != null && ownBotUserId == targetBotUserId;
+        return ownBotUserId != null && ownBotUserId.longValue() == targetBotUserId;
     }
 }
